Hopefully resolved the display issue
Oh boy did we try a number of things.  Seems as though the MAX7219 chips
that we generally buy from China are (big surprise) fakes.  This
actually wasn't problem, but provided hours of interestiner reading and
debugging attempts and deadends.
Turns out, wasn't hardware (tried/swapped) bunches of things.
Wasn't ESP8266 and its SPI impl, or interrupts, n-such.

Was the old library I was using, MaxMatrix.  I was using it before in
iClockv1 running on an Arduino (328).  It was bitbanging, but working
just fine.  Moving it to ESP8266 was easy so I just kept using it.

Ending up trying, and switching to MD_MAX72xx.  Much richer, supported
variable width fonts, and worked.  Modified on of the fonts to use more
narrow numbers.  Now 12h time fits.  Needed to rotate my GENERIC_HW
display, but otherwise all good so far.  Running for a few hours now,
WAY longer than the old library would handle.  The old approach worked
for for 2 modules (which I had for convience for some time), but
switching to 4 brought headache.