add base function for turning the display off
Add the web controlled ability to turn off the display.

If the MPR121 sesnse button is pressed, the display turns on for 30
seconds, the turns off again.

TODO
Need to add the ability to set "Night/Day" time parameters to
automatically control turning the display Off/On as well as manual
control.

The setting need to be persisted as well in SPIFFS