still chasing bad MAX7219 signals
Removed attempts to re-init frequently to reset failure.  Still think
that's kinda lame, but should work.  The bad modules act good again
without needing a power reset.  Maybe all are needed, maybe re-init of
wire, maybe data in the ESP is corrupt, ...

As it is, I moved to a bitbashing method instead of hardware SPI, and
shortened more wires.  Haven't see the problem in days....