added day-night on-off control, can now persist settings to SPIFFS
- added a feature to allow the user to set a night-off time and a day-on
window.  Nothing really enforces a "window", but you can set the
off-time to be like 23:00 (11PM) and the on-time to be 07:00 (7AM).
During that time the display will turn off.  If the sense button is
touched the display turns on for 30 seconds, then returns to off.  If
the physical button is pressed the weather is shown, then off again.
- this is controlled through the web interface
- this is now persisted to SPIFFS using the ArduinoJson library.
- other items are persisted as well, RBG settings, all display settings,
etc.
- adjusted all MAX72xx prints to only print if the display is on.  So
once the display is off, all non-requested messages are supressed, even
boot messages.  We'll see how that goes.