change json serialization to handle newly added objects
now that the config file is being use, we'll need to add new data
elements to be persisted.

Now as we read an existing file, deserialize it, and extract elements we
handle elements that don't exist.  If the elements isNull we internally
set a default.  As always, when values are changed the whole set is
serialized out.

Will still need to handle major changes were old elements may exist but
are not compatible with the new element version.